refactor(journal): break slug page into components
